+++ pp/__init__.py
@@ -120,7 +120,7 @@ __all__ = [
     "component_from_yaml",
     "units",
 ]
-__version__ = "2.2.3"
+__version__ = "2.2.4"
 
 
 if __name__ == "__main__":