+++ pp/test/test_netlist_with_routes.py
@@ -13,14 +13,16 @@ def test_netlist_with_routes():
 
     routes = pp.routing.connect_bundle(w.ports["E0"], b.ports["W0"])
     c.add(routes)
+    n = c.get_netlist()
+    connections = n["connections"]
 
     # print(routes[0].get_settings())
-    print(c.get_netlist().connections)
-    print(c.get_netlist().instances)
-    print(len(c.get_netlist().connections))
+    # print(c.get_netlist().connections)
+    # print(c.get_netlist().instances)
+    # print(len(c.get_netlist().connections))
 
     assert len(c.get_dependencies()) == 3
-    assert len(c.get_netlist().connections) == 2  # 2 components + 1 flat netlist
+    assert len(connections) == 2  # 2 components + 1 flat netlist
     return c